Apply patch
git-svn-id: svn+ssh://src.dsl.bbn.com/svn/openmap/trunk@2176 f68c23cc-7f87-4fbc-acfa-d9760b90acc1

+++ src/openmap/com/bbn/openmap/layer/e00/TX7.java
@@ -199,8 +199,8 @@ public class TX7 extends OMGraphicAdapter {
     public boolean generate(Projection proj) {
         proj.forward(llp1, pt1);
         proj.forward(llp2, pt2);
-        int dx = pt2.x - pt1.x;
-        int dy = pt2.y - pt1.y;
+        double dx = pt2.x - pt1.x;
+        double dy = pt2.y - pt1.y;
         at.setToTranslation(pt1.x, pt1.y);
         at.rotate(Math.atan2(dy, dx) - angle, 0, 0);
         double sc = Math.sqrt(dx * dx + dy * dy);
